Wrap task submission for transfer into tasks
Getting a future returned is quicker since the task submission is no longer in the main thread. So for example, HeadObject of downloads and copies no longer happen in the main thread which makes enqueueing of tasks much faster. Also, will make it easier for when we add the functionality to upload/download to and from anonseekable stream.