Add the ability to limit in memory upload chunks
This is currently helpful for multipart seekable files where we have to store
the data into memory. This will allow users to limit the number of these
partitioned chunks of data can be in memory at a time. This will also be needed
for nonseekable uploads and possibly nonseekable downloads.