Update stream chunksize for io writes
Consider the case where we have a fast network and are bounded more by io.
Given that our default io queue size is a thousand and the default io chunk
size is 8KB, then the most amount of data that can be stored in the queue is
around 8 MB. However, this is a problem because each downloaded part is 8 MB
as well so if the threads are downloading it really quick there is a
bottleneck in the queue where the queue may fill up because the enqueuing
is much quicker than the io thread is processing them. We need to bump up the
value of the io chunksize or io queue size to something closer to the
value of (max_request_threads * request_chunksize). The best performing
modification was to increase the size of the io chunks from 8KB to 64KB.