Add support for non-seekable streaming uploads
The tricky part of this change is that existing code relies on the
ability to get the length of the object being uploaded, which is
not possible for non-seekable objects. To work around that issue,
this will read data into memory up to the configured multipart
threshold to determine if it requires a multipart upload. That data
will then be used to generate the initial parts.