Fix issue with tests and added more logging
The issue was that the KeyboardInterrupt would be called during the
execution of the CompleteMultipartUpload and as a result the
AbortMultipartUpload would not delete the object because the multipart
upload already completed. The solution was to make the sleep time shorter to
help better ensure the abort happens before the complete multipart upload
and log when KeyboardInterrupts are sent and when coordinators are cancelled.