Remove resetting of associated future set
The resetting of the future set was not needed and was actually causing issues
in the done callbacks attached to removing associated futures. The reason it
was there in the first place was because we were not adding done callbacks
to remove a future's association before this was added.

Take the case if the task is the final task. If the task is final it
will completely reset the set of associated futures, but a done
callback would have been attached to the future to complete after that
task is complete. The problem is the task will first wipe out its
associated future set and then try to remove itself from the set,
which will cause an error in the callback because the set is empty.