Fix issue when uploading a previos seeked file
The current position was not being taken into account when providing
the full file size thus the ReadFileChunk wrapper was using an improper
size when calculating length and how much more is needed to be read. Note
that this does not affect the multipart uploads because the original file
object is not used and instead data is chunked out to separate file objects
from the starting position, but a sanity test was added just in case.