Add AggregatedProgressCallback
The on_progress callbacks for uploads are too grainular (like 8KB). This
is a problem for faster network environments because if you have any
synchronization mechanisms in place for on_progress, it will slow down the
transfer substantially because the read() is often much faster than the
mechanisms that may be getting invoked in the callback. The abstraction
aggregates these smaller progress updates into a larger one to help ensure
that these slower mechanisms are not called too often. The default of 256KB
was chosen as it is still pretty grainular for non-multipart uploads and
there is not too much of improvement past that value based off of testing.