Add ExecutorFuture
It essentially wraps around concurrent.futures.Future. It allows
us to do things. First and most immediate, __repr__ of Future uses
a lock and that can cause deadlocks with logging if you are logging
the object. Second, it allows to move off of concurrent.futures in
the future since the rest of this codebase now relies on
ExecutorFuture.