Add validation in state transition
This covers an edge case where a transfer gets cancelled in the _main()
of a submission task, the submission task may still have the cancelled
state overriden if it tries to set it to queued or progress. By propogating
an error, it will prevent this from happening and will actually ensure
the submission task is shutdown faster.