Create own version of ThreadPoolExecutor
The reason this was done was that concurrent.futures version was too
complicated for what was needed and because of that there was a fair
amount of undesired deadlocks from features that we do not even care
about.

This version fo ThreadPoolExecutor is much simpler in the sense that
it spins up all of its threads at once, only has submit(), does not
concern itself with waiters, and does not register any atexit callbacks.