Remove unnecessary lock
The lock can be removed because it is guaranteed that the done
state along with the result/error before the done event is
announced. Having this lock around started causing a deadlock in
the case where the main thread would call cancel, the task had
not started so done gets announced, when done gets announced it
runs the done callbacks, and if one of these done callbacks used
the result() it would cause a deadlock where cancel() already had
the lock and result() was unnecessarily requiring the lock. I also
considered switching it to an RLock, but it seemed like we should
not be introducing if it is not needed as of now.