Avoid io executor for nonranged downloads
The io executor and the other primatives that came up with it  were
unnecessary for nonranged files as there is only ever one thread working on
a particular nonranged download. These were relatively pretty heavy weight
so by not using the io executor and writing to the file-like object
immediately upon download improves both speed and memory.