Increase download test file size
This increases the size of the object we're uploading / downloading
in an integration test. The test's purpose is to verify exceptions
will cancel downloads, but the switch to sigv4 has made downloads a
bit faster. Thus the size of the download needed to be increased to
account for that. This issue was also seen in the CLI tests.