Reduce chance of race condition in tests
When adding a done callback to a future, it will execute immediately
*in the current thread* if the future is marked as done. This means
that you can have callbacks occuring simultaneously. We had a test
that was relying on a pair of callbacks being ran one at a time and
in order. When it ran into this race condition, it would fail. This
adds a delay to the completion of the future so that the race condition
is less likely.