Fix flaky integration test
The test was trying to time when a transfer started.  If it sends
a ctrl-c before or after the test starts, we don't see the keyboard
interrupt.  Previously the test did this by sleeping for some amount
of time.

This commit replaces the sleep with an event.  We use the s3transfer
subscribers to wait until we get a progress callback, then we trigger
an event that causes the main thread to raise the KeyboardInterrupt.