Make ctrl-c upload test more robust
This adds a similar fix to the ctrl-c download test where we
measure the time from when we ctrl-c to when it exits, which is
what we really want to test.

I also updated the tests to run quicker.  An earlier commit
switched the object_exists call to use a waiter, but did not
verify the slower runtime of the tests in the negative case.
Rather than wait for the waiter to timeout, I've added an
object_not_exists.  This cuts the time to run this test down by
an order or magnitude.  A subsequent commit will fix the remaining
cases.