Fix test_download_futures_fail_triggers_shutdown with Python 3.8
The behaviour of set_exception for futures changed in Python 3.8, it'll now
raise concurrent.futures.InvalidStateError if the future is already done [1],
which is the case in this test because set_result has already been called on
the future.

1: https://bugs.python.org/issue33238

Fix the test by not using the future from SequentialExecutor, and instead
creating a future which doesn't have a result.