Add CRT process lock utility
The CRT S3 client performs best when there is only one
instance of it running on a host. This lock allows an application to
signal across processes whether there is another process of the same
application using the CRT S3 client and prevent spawning more than one
CRT S3 clients running on the system for that application.