Update throughput target defaults
Instead of defaulting to five gigabits per second when no target
througput is configured, s3transfer will use the CRT to determine
if there is a recommended target throughput to use. If it is unable,
to determine a recommended target throughput, it will default to
ten gigabits per second, which is the current throughput default
for the CRT S3 client.

This behavior was updated to help minimize the amount of configuration
required to get the fastest throughputs from the CRT transfer client.
The fallback throughput was updated from five to ten gigabits so
that the CRT S3 integration is consistent with the underlying CRT S3
client's defaults.